+++ esmf_regrid/__init__.py
@@ -11,4 +11,5 @@ except ImportError as exc:
 from .constants import Constants, check_method, check_norm
 from .schemes import *
 
-__version__ = "0.8.dev0"
+
+__version__ = "0.10.dev0"