+++ esmf_regrid/tests/unit/experimental/unstructured_scheme/test_MeshToGridESMFRegridder.py
@@ -68,11 +68,11 @@ def test_flat_cubes():
     expected_cube.data = result.data
     assert expected_cube == result
 
-
+@pytest.mark.parametrize("nsi", [0, 1])
 @pytest.mark.parametrize(
     "method", [Constants.Method.BILINEAR, Constants.Method.NEAREST]
 )
-def test_node_friendly_methods(method):
+def test_node_friendly_methods(method, nsi):
     """
     Basic test for :class:`esmf_regrid.experimental.unstructured_scheme.MeshToGridESMFRegridder`.
 
@@ -83,8 +83,8 @@ def test_node_friendly_methods(method):
     lon_bounds = (-180, 180)
     lat_bounds = (-90, 90)
     tgt = _grid_cube(n_lons, n_lats, lon_bounds, lat_bounds, circular=True)
-    face_src = _gridlike_mesh_cube(n_lons, n_lats, location="face")
-    node_src = _gridlike_mesh_cube(n_lons, n_lats, location="node")
+    face_src = _gridlike_mesh_cube(n_lons, n_lats, location="face", nsi=nsi)
+    node_src = _gridlike_mesh_cube(n_lons, n_lats, location="node", nsi=nsi)
 
     face_src = _add_metadata(face_src)
     node_src = _add_metadata(node_src)