+++ esmf_regrid/tests/unit/schemes/test_ESMFNearest.py
@@ -8,15 +8,24 @@ from esmf_regrid.schemes import ESMFNearest
 from esmf_regrid.tests.unit.schemes.__init__ import (
     _test_mask_from_init,
     _test_mask_from_regridder,
+    _test_non_degree_crs,
 )
 from esmf_regrid.tests.unit.schemes.test__cube_to_GridInfo import _grid_cube
 from esmf_regrid.tests.unit.schemes.test__mesh_to_MeshInfo import (
+    _gridlike_mesh,
     _gridlike_mesh_cube,
 )
 
 
 @pytest.mark.parametrize(
-    "src_type,tgt_type", [("grid", "grid"), ("grid", "mesh"), ("mesh", "grid")]
+    "src_type,tgt_type",
+    [
+        ("grid", "grid"),
+        ("grid", "mesh"),
+        ("grid", "just_mesh"),
+        ("mesh", "grid"),
+        ("mesh", "mesh"),
+    ],
 )
 def test_cube_regrid(src_type, tgt_type):
     """
@@ -24,7 +33,10 @@ def test_cube_regrid(src_type, tgt_type):
 
     Checks that regridding occurs.
     """
-    scheme_default = ESMFNearest()
+    if tgt_type == "just_mesh":
+        scheme_default = ESMFNearest(tgt_location="face")
+    else:
+        scheme_default = ESMFNearest()
 
     n_lons_src = 6
     n_lons_tgt = 3
@@ -47,8 +59,13 @@ def test_cube_regrid(src_type, tgt_type):
         expected_data_default = np.zeros([n_lats_tgt, n_lons_tgt])
         expected_mask = np.zeros([n_lats_tgt, n_lons_tgt])
         expected_mask[0, 0] = 1
-    else:
+    elif tgt_type == "mesh":
         tgt = _gridlike_mesh_cube(n_lons_tgt, n_lats_tgt)
+        expected_data_default = np.zeros([n_lats_tgt * n_lons_tgt])
+        expected_mask = np.zeros([n_lats_tgt * n_lons_tgt])
+        expected_mask[0] = 1
+    elif tgt_type == "just_mesh":
+        tgt = _gridlike_mesh(n_lons_tgt, n_lats_tgt)
         expected_data_default = np.zeros([n_lats_tgt * n_lons_tgt])
         expected_mask = np.zeros([n_lats_tgt * n_lons_tgt])
         expected_mask[0] = 1
@@ -57,7 +74,10 @@ def test_cube_regrid(src_type, tgt_type):
 
     result_default = src.regrid(tgt, scheme_default)
 
-    expected_cube_default = tgt.copy()
+    if tgt_type == "just_mesh":
+        expected_cube_default = _gridlike_mesh_cube(n_lons_tgt, n_lats_tgt)
+    else:
+        expected_cube_default = tgt.copy()
     expected_cube_default.data = expected_data_default
 
     assert expected_cube_default == result_default
@@ -81,3 +101,8 @@ def test_mask_from_regridder(mask_keyword):
     Checks that use_src_mask and use_tgt_mask are passed down correctly.
     """
     _test_mask_from_regridder(ESMFNearest, mask_keyword)
+
+
+def test_non_degree_crs():
+    """Test for coordinates with non-degree units."""
+    _test_non_degree_crs(ESMFNearest)